+++ packages/nutui/components/cellgroup/cellgroup.vue
@@ -1,7 +1,7 @@
-<script setup lang="ts">
+<script lang="ts" setup>
 import { computed, defineComponent, useSlots } from 'vue'
-import { getMainClass } from '../_utils'
 import { PREFIX } from '../_constants'
+import { getMainClass } from '../_utils'
 import { cellgroupProps } from './cellgroup'
 
 const props = defineProps(cellgroupProps)
@@ -53,5 +53,5 @@ export default defineComponent({
 </template>
 
 <style lang="scss">
-@import './index';
+@import "./index";
 </style>