+++ packages/nutui/components/comment/components/CmtImages.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import type { PropType } from 'vue'
 import { defineComponent, onMounted, ref, watch } from 'vue'
-import NutIcon from '../../icon/icon.vue'
 import { PREFIX } from '../../_constants'
+import NutIcon from '../../icon/icon.vue'
 
 const props = defineProps({
   type: {
@@ -94,7 +94,7 @@ export default defineComponent({
       <view class="nut-comment-images__play" />
     </view>
     <!-- images -->
-    <template v-for="(itI, index) in images" :key="itI.id">
+    <template v-for="(itI, index) in images" :key="index">
       <view
         v-if="(type === 'multi' && videos.length + index < 9) || type !== 'multi'"
         class="nut-comment-images__item nut-comment-images__item--imgbox"