+++ packages/nutui/components/composables.ts
@@ -1,1 +1,2 @@
-export * from './toast/use-toast'
+export { useNotify } from './notify/use-notify'
+export { useToast } from './toast/use-toast'